Refactor the UI to be more modern and user-friendly. This includes restructuring the project, refactoring components to use CSS modules, and improving the overall layout and styling.
