+++ package.json
@@ -8,9 +8,11 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^13.5.0",
+    "classnames": "^2.5.1",
     "lucide-react": "^0.482.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.7.0",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
