+++ src/layouts/MainLayout.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import Header from '../components/Header';
+import Footer from '../components/Footer';
+
+const MainLayout = ({ children }) => {
+  return (
+    <div className="main-layout">
+      <Header />
+      <main>{children}</main>
+      <Footer />
+    </div>
+  );
+};
+
+export default MainLayout;
