+++ src/pages/StockChartAnalyzer.test.js
